Actualización modulo de programación de cafe Actualización del modelo de la base de datos Cambio en la UI de login y programación de cafe

+++ src/siscafe/persistence/NavyAgentJpaController.java
@@ -6,16 +6,18 @@
 package siscafe.persistence;
 
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import siscafe.model.PackagingCaffee;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
+import siscafe.controller.exceptions.IllegalOrphanException;
+import siscafe.controller.exceptions.NonexistentEntityException;
 import siscafe.model.NavyAgent;
-import siscafe.persistence.exceptions.NonexistentEntityException;
-import siscafe.persistence.exceptions.PreexistingEntityException;
 
 /**
  *
@@ -32,18 +34,31 @@ public class NavyAgentJpaController implements Serializable {
         return emf.createEntityManager();
     }
 
-    public void create(NavyAgent navyAgent) throws PreexistingEntityException, Exception {
+    public void create(NavyAgent navyAgent) {
+        if (navyAgent.getPackagingCaffeeList() == null) {
+            navyAgent.setPackagingCaffeeList(new ArrayList<PackagingCaffee>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            em.persist(navyAgent);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            if (findNavyAgent(navyAgent.getId()) != null) {
-                throw new PreexistingEntityException("NavyAgent " + navyAgent + " already exists.", ex);
+            List<PackagingCaffee> attachedPackagingCaffeeList = new ArrayList<PackagingCaffee>();
+            for (PackagingCaffee packagingCaffeeListPackagingCaffeeToAttach : navyAgent.getPackagingCaffeeList()) {
+                packagingCaffeeListPackagingCaffeeToAttach = em.getReference(packagingCaffeeListPackagingCaffeeToAttach.getClass(), packagingCaffeeListPackagingCaffeeToAttach.getId());
+                attachedPackagingCaffeeList.add(packagingCaffeeListPackagingCaffeeToAttach);
             }
-            throw ex;
+            navyAgent.setPackagingCaffeeList(attachedPackagingCaffeeList);
+            em.persist(navyAgent);
+            for (PackagingCaffee packagingCaffeeListPackagingCaffee : navyAgent.getPackagingCaffeeList()) {
+                NavyAgent oldNavyAgentIdOfPackagingCaffeeListPackagingCaffee = packagingCaffeeListPackagingCaffee.getNavyAgentId();
+                packagingCaffeeListPackagingCaffee.setNavyAgentId(navyAgent);
+                packagingCaffeeListPackagingCaffee = em.merge(packagingCaffeeListPackagingCaffee);
+                if (oldNavyAgentIdOfPackagingCaffeeListPackagingCaffee != null) {
+                    oldNavyAgentIdOfPackagingCaffeeListPackagingCaffee.getPackagingCaffeeList().remove(packagingCaffeeListPackagingCaffee);
+                    oldNavyAgentIdOfPackagingCaffeeListPackagingCaffee = em.merge(oldNavyAgentIdOfPackagingCaffeeListPackagingCaffee);
+                }
+            }
+            em.getTransaction().commit();
         } finally {
             if (em != null) {
                 em.close();
@@ -51,12 +66,45 @@ public class NavyAgentJpaController implements Serializable {
         }
     }
 
-    public void edit(NavyAgent navyAgent) throws NonexistentEntityException, Exception {
+    public void edit(NavyAgent navyAgent) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            NavyAgent persistentNavyAgent = em.find(NavyAgent.class, navyAgent.getId());
+            List<PackagingCaffee> packagingCaffeeListOld = persistentNavyAgent.getPackagingCaffeeList();
+            List<PackagingCaffee> packagingCaffeeListNew = navyAgent.getPackagingCaffeeList();
+            List<String> illegalOrphanMessages = null;
+            for (PackagingCaffee packagingCaffeeListOldPackagingCaffee : packagingCaffeeListOld) {
+                if (!packagingCaffeeListNew.contains(packagingCaffeeListOldPackagingCaffee)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain PackagingCaffee " + packagingCaffeeListOldPackagingCaffee + " since its navyAgentId field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<PackagingCaffee> attachedPackagingCaffeeListNew = new ArrayList<PackagingCaffee>();
+            for (PackagingCaffee packagingCaffeeListNewPackagingCaffeeToAttach : packagingCaffeeListNew) {
+                packagingCaffeeListNewPackagingCaffeeToAttach = em.getReference(packagingCaffeeListNewPackagingCaffeeToAttach.getClass(), packagingCaffeeListNewPackagingCaffeeToAttach.getId());
+                attachedPackagingCaffeeListNew.add(packagingCaffeeListNewPackagingCaffeeToAttach);
+            }
+            packagingCaffeeListNew = attachedPackagingCaffeeListNew;
+            navyAgent.setPackagingCaffeeList(packagingCaffeeListNew);
             navyAgent = em.merge(navyAgent);
+            for (PackagingCaffee packagingCaffeeListNewPackagingCaffee : packagingCaffeeListNew) {
+                if (!packagingCaffeeListOld.contains(packagingCaffeeListNewPackagingCaffee)) {
+                    NavyAgent oldNavyAgentIdOfPackagingCaffeeListNewPackagingCaffee = packagingCaffeeListNewPackagingCaffee.getNavyAgentId();
+                    packagingCaffeeListNewPackagingCaffee.setNavyAgentId(navyAgent);
+                    packagingCaffeeListNewPackagingCaffee = em.merge(packagingCaffeeListNewPackagingCaffee);
+                    if (oldNavyAgentIdOfPackagingCaffeeListNewPackagingCaffee != null && !oldNavyAgentIdOfPackagingCaffeeListNewPackagingCaffee.equals(navyAgent)) {
+                        oldNavyAgentIdOfPackagingCaffeeListNewPackagingCaffee.getPackagingCaffeeList().remove(packagingCaffeeListNewPackagingCaffee);
+                        oldNavyAgentIdOfPackagingCaffeeListNewPackagingCaffee = em.merge(oldNavyAgentIdOfPackagingCaffeeListNewPackagingCaffee);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -74,7 +122,7 @@ public class NavyAgentJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -85,6 +133,17 @@ public class NavyAgentJpaController implements Serializable {
                 navyAgent.getId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The navyAgent with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<PackagingCaffee> packagingCaffeeListOrphanCheck = navyAgent.getPackagingCaffeeList();
+            for (PackagingCaffee packagingCaffeeListOrphanCheckPackagingCaffee : packagingCaffeeListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This NavyAgent (" + navyAgent + ") cannot be destroyed since the PackagingCaffee " + packagingCaffeeListOrphanCheckPackagingCaffee + " in its packagingCaffeeList field has a non-nullable navyAgentId field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(navyAgent);
             em.getTransaction().commit();
